- fixed cpu usage statistic for Linux, now correct values are delivered for the relative values of cpu usage

+++ HomerBase/include/HBThread.h
@@ -64,7 +64,7 @@ public:
     static int GetPId();
     static int GetPPId();
     static std::vector<int> GetTIds();
-    static bool GetThreadStatistic(int pTid, int &pMemVirtual, int &pMemPhysical, int &pPid, int &pPPid, float &pLoadUser, float &pLoadSystem, float &pLoadTotal, int &pPriority, int &pNice, int &pThreadCount, unsigned long long &pLastUserTicsThread, unsigned long long &pLastKernelTicsThread, unsigned long long &pLastUserTicsSystem, unsigned long long &pLastKernelTicsSystem);
+    static bool GetThreadStatistic(int pTid, unsigned long &pMemVirtual, unsigned long &pMemPhysical, int &pPid, int &pPPid, float &pLoadUser, float &pLoadSystem, float &pLoadTotal, int &pPriority, int &pNice, int &pThreadCount, unsigned long long &pLastUserTicsThread, unsigned long long &pLastKernelTicsThread, unsigned long long &pLastUserTicsSystem, unsigned long long &pLastKernelTicsSystem);
 
 private:
     static void* StartThreadStaticWrapperUniversal(void* pThread);
